completed/01_arXiv_논문_요약_에이전트/requirements.txt
feedparser>=6.0.11
anthropic>=0.40.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
tenacity>=8.2.3
loguru>=0.7.2
